Fix autoloading for PHPUnit-3.6.
Originally we had a line like "if (!spl_autoload_functions()) {
spl_autoload_register(...); }" in the autoload setup for the test
suites of the components. This was unified in Horde/Test/Autoload.php
at some point and later modified so that the PHPUnit specific
autoloader (3.5) would be ignored. With PHPUnit-3.6 we get a slew of
additional autoloaders which invalidated that method.

I could have added the new autoloaders to the ignore list but that
might as well fail with 3.7 again. So I removed the "if
(!spl_autoload_functions())" check instead. I didn't see a valid
reason for that check anymore. The additional autoloader won't hurt
even if we required Horde/Autoloader/Default.php (or any other
autoloader capable of loading the Horde classes) before pulling
Horde/Test/Autoload.php.

Please shout if you can think of a reason why that check might be
valid.

I only added a define() directive as a safeguard in case
Horde/Test/Autoload.php is not just required once.

In principle the test suites should work with 3.6 now. There are still
a few incompatibilites that I noted though:

1) assertType() needs to be assertInternalType() now

2) I will add the PHPUnit_Story dependency to Horde_Test as we have a
   few story based test - though I want to get rid of those in the
   long term.

3) Some constraint interfaces don't seem to match anymore
   (Kolab_Server) and need to be adapted to avoid warnings.